Project Files/sustainable-smart-city-assistant/app/requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
sentence-transformers==2.2.2
pinecone-client==2.2.4
scikit-learn==1.3.2
pandas==2.1.3
numpy
matplotlib==3.8.2
seaborn==0.13.0
streamlit-option-menu==0.3.6
ibm-watson-machine-learning
